requirements.txt
pyYaml
bs4
requests
python-bidi
textual
pytest
coverage
